불필요한 파일 정리
